* Return u64 from MemoryArea::size
* Add comments clarifying truncation of physical XSDT address on x86

+++ src/memory_map.rs
@@ -38,8 +38,8 @@ impl MemoryArea {
         (self.base_addr + self.length)
     }
 
-    pub fn size(&self) -> usize {
-        self.length as usize
+    pub fn size(&self) -> u64 {
+        self.length
     }
 }
